requirements.txt
torch
torchvision
torchaudio
matplotlib
jupyterlab
scikit-learn
onnx
onnxruntime
flask
tensorboard
numpy
Pillow